Move logger to internal/ — QNTX-specific, not for external import
