+++ cmd/qntx/commands/ix.go
@@ -7,7 +7,7 @@ import (
 	"github.com/teranos/QNTX/am"
 	"github.com/teranos/QNTX/db"
 	"github.com/teranos/QNTX/errors"
-	"github.com/teranos/QNTX/logger"
+	"github.com/teranos/QNTX/internal/logger"
 	"github.com/teranos/QNTX/pulse/async"
 	"github.com/teranos/QNTX/sym"
 )
