+++ internal/logger/minimal_encoder.go
