+++ internal/logger/minimal_encoder_test.go
