+++ internal/logger/symbol.go
