+++ plugin/grpc/banner.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/teranos/QNTX/logger"
+	"github.com/teranos/QNTX/internal/logger"
 	"go.uber.org/zap"
 )
