+++ pulse/async/worker.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/teranos/QNTX/am"
 	"github.com/teranos/QNTX/errors"
-	"github.com/teranos/QNTX/logger"
+	"github.com/teranos/QNTX/internal/logger"
 	"github.com/teranos/QNTX/pulse/budget"
 	"go.uber.org/zap"
 )
