+++ server/broadcast.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/teranos/QNTX/errors"
 	"github.com/teranos/QNTX/graph"
-	"github.com/teranos/QNTX/logger"
+	"github.com/teranos/QNTX/internal/logger"
 	"github.com/teranos/QNTX/pulse/async"
 	"github.com/teranos/QNTX/pulse/schedule"
 	"github.com/teranos/QNTX/server/wslogs"
