+++ server/client.go
@@ -18,7 +18,7 @@ import (
 	"github.com/teranos/QNTX/errors"
 	"github.com/teranos/QNTX/graph"
 	grapherr "github.com/teranos/QNTX/graph/error"
-	"github.com/teranos/QNTX/logger"
+	"github.com/teranos/QNTX/internal/logger"
 	"github.com/teranos/QNTX/server/syscap"
 	"github.com/teranos/QNTX/server/wslogs"
 )
